app/requirements.txt
alembic==1.8.1
anyio==3.6.1
asgiref==3.5.2
attrs==22.1.0
bcrypt==4.0.1
Brotli==1.0.9
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
ConfigArgParse==1.5.3
cryptography==39.0.0
ecdsa==0.18.0
fabric==2.7.1
fake==0.8
Faker==16.1.0
fastapi==0.85.0
fastapi-keycloak==1.0.7
fastapi-pagination==0.10.0
Flask==2.2.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==22.10.2
geventhttpclient==2.0.8
greenlet==2.0.1
h11==0.14.0
httptools==0.5.0
idna==3.4
iniconfig==1.1.1
invoke==1.7.3
itsdangerous==2.1.2
Jinja2==3.1.2
locust==2.14.1
Mako==1.2.3
MarkupSafe==2.1.1
msgpack==1.0.4
packaging==21.3
paramiko==2.12.0
pathlib2==2.3.7.post1
pluggy==1.0.0
psutil==5.9.4
psycopg2==2.9.4
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
PyNaCl==1.5.0
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
roundrobin==0.0.4
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
tomli==2.0.1
typing-extensions==4.4.0
urllib3==1.26.12
uvicorn==0.18.3
watchfiles==0.17.0
websockets==10.3
Werkzeug==2.2.2
zope.event==4.6
zope.interface==5.5.2
